Move snap/divisor helper methods to inside `ControlPointInfo`
